backend/requirements.txt
flask
flask-jwt-extended
firebase-admin
pycryptodome
python-dotenv
